ver 4.6.4 Added SimplifyPolygon and SimplifyPolygons functions.
git-svn-id: svn://svn.code.sf.net/p/polyclipping/code/trunk@230 758e7f8d-410f-429b-b4f6-457a56196ea1

+++ cpp/cpp_cairo/cairo_clipper.cpp
@@ -30,7 +30,7 @@ namespace ClipperLib {
 
       void transform_point(cairo_t* pen, Transform transform, long64* x, long64* y)
       {
-        double _x = *x, _y = *y;
+        double _x = (double)*x, _y = (double)*y;
         switch (transform)
         {
           case tDeviceToUser:
